Rectification 2 dictionnaire Firefox

+++ mintpostinstall.sh
@@ -52,7 +52,12 @@ nemo -q
 
 # Installation du dictionnaire Français sous Firefox et activation
 wget -O /tmp/fr-FR.xpi "wget -O /tmp/fr-FR.xpi "https://addons.mozilla.org/firefox/downloads/latest/french-dictionary/latest.xpi""
-firefox --install-addon /tmp/french-dictionnary.xpi
+firefox --install-addon /tmp/french-dictionary.xpi
+PROFILE=$(find ~/.mozilla/firefox -maxdepth 1 -type d -name "*.default*" | head -n 1)
+if [ -z "$PROFILE" ]; then
+    echo "Profil Firefox non trouvé. Vérifie que Firefox a déjà été lancé au moins une fois."
+    exit 1
+fi
 echo 'user_pref("spellchecker.dictionary", "fr-FR");' >> "$PROFILE/prefs.js"
 rm /tmp/french-dictionary.xpi
